[DEMO] DependencyController
### What's done:
 * Moved all endpoints that are connected with files to DependencyController (/demo/internal/files)
 * Changed base path for ManagementController (/demo/internal/manager)
 * Fixed all endpoints
